Fix for DNC animations
Some abilities were set to use action type 14 when they needed to be 6.

Also fixed an issue with presto script,
was trying to add and set finishing moves when this is not the case,
as the extra step is added after a successful step is performed and was not removing the effect if a successful step was landed,
this has been corrected in the corresponding step scripts.